+++ CaledonEngine/Systems/Rendering/Shapes/Triangle.cpp
@@ -3,9 +3,10 @@
 | Author: Chandler Mays
 ------------------------------*/
 #include "Triangle.h"
-#include "CaledonEngine/Systems/Rendering/Renderer.h"
-#include "CaledonEngine/Utilities/Math/Rect.h"
-#include "CaledonEngine/Utilities/Math/Vector2.h"
+
+#include "Systems/Rendering/Renderer.h"
+#include "Utilities/Math/Rect.h"
+#include "Utilities/Math/Vector2.h"
 
 /*-----------------------------------
 | --- Public Method Definitions --- |
@@ -25,9 +26,9 @@ void CE::Triangle::Render(Renderer* pRenderer, const Rect& destRect, const Color
 	if (!pRenderer)
 		return;
 
-	Vector2i v1(destRect.m_x + destRect.m_width / 2, destRect.m_y);
-	Vector2i v2(destRect.m_x, destRect.m_y + destRect.m_height);
-	Vector2i v3(destRect.m_x + destRect.m_width, destRect.m_y + destRect.m_height);
+	Vector2f v1(destRect.m_x + destRect.m_width / 2, destRect.m_y);
+	Vector2f v2(destRect.m_x, destRect.m_y + destRect.m_height);
+	Vector2f v3(destRect.m_x + destRect.m_width, destRect.m_y + destRect.m_height);
 
 	pRenderer->DrawTriangle(v1, v2, v3, color, m_isFilled);
 }
